requirements.txt
keras
tensorflow
pillow
h5py
numpy